extract word api
